requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-log==0.4.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.21.2
dotty-dict==1.3.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==1.3.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
instagrapi==2.1.3
instaloader==4.14.1
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
more-itertools==10.7.0
multidict==6.4.3
nh3==0.2.21
packaging==25.0
pillow==11.2.1
pkginfo==1.12.1.2
postgrest==0.10.7
propcache==0.3.1
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
python-telegram-bot==20.6
pywin32-ctypes==0.2.3
readme_renderer==44.0
realtime==1.0.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tomlkit==0.13.2
tqdm==4.67.1
twine==3.8.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==12.0
wheel==0.45.1
yarl==1.20.0
yt-dlp==2025.4.30
zipp==3.21.0
